AI_BOOK/pyproject.toml
[project]
name = "aibook"
version = "0.1.0"
description = "Een modern boek over Kunstmatige Intelligentie"
authors = [
    {name = "AI Book Team"}
]
dependencies = [
    "flask>=3.0.2",
    "markdown>=3.7",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.1",
    "python-frontmatter>=1.0.0",
    "watchdog>=3.0.0"
]
requires-python = ">=3.9"

[project.scripts]
aibook = "aibook.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aibook"]
include = ["src/content/*", "src/templates/*", "src/static/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py39" 